requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# code_cvpr_9521/data/additional_transforms.py: 9
# code_cvpr_9521/data/datamgr.py: 4
# code_cvpr_9521/data/dataset.py: 4
Pillow == 8.3.1

# code_cvpr_9521/data/feature_loader.py: 3
# code_cvpr_9521/save_features.py: 7
h5py == 2.10.0

# code_cvpr_9521/backbone.py: 7
# code_cvpr_9521/data/datamgr.py: 5
# code_cvpr_9521/data/dataset.py: 6
# code_cvpr_9521/data/feature_loader.py: 2
# code_cvpr_9521/filelists/CUB/write_filelist.py: 1
# code_cvpr_9521/filelists/cifar/write_filelist.py: 1
# code_cvpr_9521/filelists/miniImagenet/write_filelist.py: 1
# code_cvpr_9521/filelists/tieredImagenet/write_filelist.py: 1
# code_cvpr_9521/io_utils.py: 1
# code_cvpr_9521/methods/baselinefinetune.py: 5
# code_cvpr_9521/methods/baselinetrain.py: 7
# code_cvpr_9521/methods/cad.py: 6
# code_cvpr_9521/methods/meta_template.py: 5
# code_cvpr_9521/methods/protonet.py: 5
# code_cvpr_9521/test.py: 4
# code_cvpr_9521/train.py: 3
# code_cvpr_9521/utils.py: 2
numpy == 1.19.2

# code_cvpr_9521/backbone.py: 3,4,5,8,9
# code_cvpr_9521/backbone_res12.py: 1,2,3
# code_cvpr_9521/data/additional_transforms.py: 8
# code_cvpr_9521/data/datamgr.py: 3
# code_cvpr_9521/data/dataset.py: 3
# code_cvpr_9521/data/feature_loader.py: 1
# code_cvpr_9521/methods/baselinefinetune.py: 2,3,4,6
# code_cvpr_9521/methods/baselinetrain.py: 4,5,6,8
# code_cvpr_9521/methods/cad.py: 3,4,5,7
# code_cvpr_9521/methods/meta_template.py: 2,3,4,6
# code_cvpr_9521/methods/protonet.py: 2,3,4,6
# code_cvpr_9521/save_features.py: 4,5,9
# code_cvpr_9521/test.py: 3,5,6,7,9
# code_cvpr_9521/train.py: 4,5,6,7,8
# code_cvpr_9521/utils.py: 1
torch == 1.7.1

# code_cvpr_9521/data/datamgr.py: 6
# code_cvpr_9521/data/dataset.py: 7
torchvision == 0.8.2

# code_cvpr_9521/test.py: 27
# code_cvpr_9521/train.py: 25
tqdm == 4.62.2
